Gpu channel: Declare `release_count` for DeferredRequest.
This CL updates the gpu_channel.mojom definition so that
`release_count` becomes a top-level field for DeferredRequest.
Correspondingly, it removes redundant `release_id` from sub-fields.

Previously `release_id` is defined as uint32_t, which is inconsistent
with SyncPointManager/SyncTokens. This CL changes to use uint64_t.

This CL is one step towards maintaining more complete
sync point dependency information for dependency analysis.

Please see the design for more details at
https://docs.google.com/document/d/1OF6OQnNljnH4v6bmg24gkuNj3zfsRAU47wONcTkHk1I

Bug: b/324276400
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5512935
Commit-Queue: Yuzhu Shen <yzshen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1297818}

+++ gpu/ipc/client/shared_image_interface_proxy.h
@@ -143,7 +143,7 @@ class SharedImageInterfaceProxy {
   const raw_ptr<GpuChannelHost> host_;
   const int32_t route_id_;
   base::Lock lock_;
-  uint32_t next_release_id_ GUARDED_BY(lock_) = 0;
+  uint64_t next_release_id_ GUARDED_BY(lock_) = 0;
   uint32_t last_flush_id_ GUARDED_BY(lock_) = 0;
 
   // A buffer used to upload initial data during SharedImage creation.
